requirements.test.txt
# This should match the version in custom_components/auroraplus/manifest.json
auroraplus@git+https://github.com/LeighCurran/AuroraPlus@main
homeassistant_historical_sensor

# Test dependencies
pytest
pytest-cov
pytest-homeassistant-custom-component
pytest-ruff
